scraper/requirements.txt
fastapi
uvicorn
scrapy
scrapy-playwright
requests
pydantic
python-dotenv
loguru
